+++ tools_proj/XCPP/GameCore/QteStatusList.cpp
@@ -37,7 +37,7 @@ void QteStatusList::ReadTable()
 void QteStatusList::GetRow(int idx,QteStatusListRow* row)
 {
 	size_t len = m_data.size();
-	if(idx<len)
+	if(idx<(int)len)
 	{
 		*row = m_data[idx];
 	}
