+++ tools_proj/XCPP/GameCore/XEntityPresentation.cpp
@@ -114,7 +114,7 @@ void XEntityPresentation::ReadTable()
 void XEntityPresentation::GetRow(int idx,XEntityPresentationRow* row)
 {
 	size_t len = m_data.size();
-	if(idx<len)
+	if(idx<(int)len)
 	{
 		*row = m_data[idx];
 	}
